refactor : portal 디렉토리 생성 및 관련 컴포넌트 이동
