+++ config-overrides.js
@@ -11,6 +11,7 @@ module.exports = override(
     '@data': path.resolve(__dirname, 'src/data'),
     '@hooks': path.resolve(__dirname, 'src/hooks'),
     '@pages': path.resolve(__dirname, 'src/pages'),
+    '@portals': path.resolve(__dirname, 'src/portals'),
     '@query': path.resolve(__dirname, 'src/query'),
     '@router': path.resolve(__dirname, 'src/router'),
     '@socket': path.resolve(__dirname, 'src/socket'),
