+++ src/App.tsx
@@ -1,13 +1,13 @@
 import { useRoutes } from 'react-router-dom';
 
-import AddressPortal from '@components/AddressPortal';
-import KeywordAddressPortal from '@components/KeywordAddressPortal';
-import LoginPortal from '@components/LoginPortal';
 import NavigationBar from '@components/NavigationBar';
-import OptionPortal from '@components/OptionPortal';
-import SearchPortal from '@components/SearchPortal';
 import useCheckIsLogin from '@hooks/useCheckIsLogin';
 import useNoticeAlarm from '@hooks/useNoticeAlarm';
+import AddressPortal from '@portals/AddressPortal';
+import KeywordAddressPortal from '@portals/KeywordAddressPortal';
+import LoginPortal from '@portals/LoginPortal';
+import OptionPortal from '@portals/OptionPortal';
+import SearchPortal from '@portals/SearchPortal';
 import { routes } from '@router';
 import Styles from '@styles';
