+++ src/components/SideBar/index.tsx
@@ -1,9 +1,9 @@
 import { Link, useLocation } from 'react-router-dom';
 import { useSetRecoilState } from 'recoil';
 
-import Portal from '@components/Portal';
 import * as S from '@components/SideBar/SideBar.style';
 import { sideBarContentsInfo } from '@components/SideBar/SideBarContents';
+import Portal from '@portals/Portal';
 import { PortalNameType, portalState } from '@store/portal';
 const portalName: PortalNameType = 'sidebar';
