+++ src/portals/SearchPortal/Search.style.tsx
