+++ tsconfig.path.json
@@ -12,6 +12,7 @@
       "@data/*": ["data/*"],
       "@hooks/*": ["hooks/*"],
       "@pages/*": ["pages/*"],
+      "@portals/*": ["portals/*"],
       "@query/*": ["query/*"],
       "@router": ["router"],
       "@router/*": ["router/*"],
